+++ docs/devel/html/navtreedata.js
@@ -53,7 +53,7 @@ var NAVTREEINDEX =
 [
 "annotated.html",
 "struct_p16_emap__t.html#a0d1b31ab7601e1be181b1a1fc3501dbb",
-"struct_p18map__t.html#aa93099bb5441da33980faac01c78c966"
+"struct_p18map__t.html#a83190248266d625affee609e1ce656d9"
 ];
 
 var SYNCONMSG = 'click to disable panel synchronisation';